requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
drf-yasg==1.21.7
drf-spectacular==0.27.2
django-cors-headers==4.4.0
sentry-sdk==2.11.0
